Modularize data storage
